Initial commit
This SourceMod extension wraps Jansson, a C library for encoding,
decoding and manipulating JSON data. Jansson is licensed under the
MIT license.
It's full text can be read here:
http://www.opensource.org/licenses/mit-license.php

Details and documentation to the library can be found here:
http://www.digip.org/jansson/

I've adapted parts of the documentation for the pawn language
and recommend referring to smjansson.inc for SourceMod specific
details. Still pretty much of it is copy'n'pasted.
Thanks to akheron for providing such detailed documentation
and a nice library in the first place.

As this is based off the 2.3.1 sources currently, I cherry picked
some of the younger commits to allow building on windows.
Looking forward to the next official release.

The original sources can be found here:
https://github.com/akheron/jansson